refactor: organize test utils
